Rebuild the build configuration from scratch
Migrate from scala classes extending `Build` to sbt files directly. SBT
will deprecate its use soon, so it's better to be prepared as soon as
possible. Also, add a task to the spores-core project that writes the
classpath with the compile-time dependencies into a file.

This extra step is done to fix `ClassNotFoundException` in neg tests,
because the classpath for `Toolbox` only uses the compiled spore classes.