Fix #52: top-level classes are static
This is an important issue. Top-level classes and traits seem to get the static flag because they are indeed static from a denotation point of view. They are not static as in `the static keyword`.

It seems this will be a recurrent problem in the future, spores play around with symbols and scalac handles them in weird ways (and some bugs are already filed for symbol assignation).

Further, it seems that neg tests are not handled appropriately and some issues are not reproducible. This issue wasn't reproducible until I removed the neg test and put the content in an independent scala file. Then, the issue appeared. I'll fix this in an upcoming commit.