Add `assumeClosed` annotation
`assumeClosed` is very handy for those dealing with binaries whose class hierarchies cannot be fully checked. It provides a escape hatch for those intrepid users who trust software from third-parties. Users that use this need to ensure that the class hierarchy is not being extended somewhere else, by them or by another jar in the classpath.