Fix dependency issues in compiler plugin
Compiler plugins are not meant to depend on external jars. However, ours depends on spores and several other libraries for developer productivity (fansi and sourcecode). Therefore, spores-serialization needs to force the user to add the spores macro library to their dependencies, so that they are available in the classpath and spores subtype checks and annotation type checks can be performed.

This commit introduces an elegant fix that always assumes that spores depends on fansi and sourcecode, and so they bring the only three dependencies to the plugin for free. This is a better approach that assembling everything in a jar, which would be awful because dependency versions would be fixed and incompatibilities could happen at userland.