Propagate concrete types for the CHA analysis
This change is prone to changes, but works so far. It makes sure that the concrete type from the spore definition site is propagated through all the analysis, so that type parameters with different type names in the same class hierarchy can be correctly mapped. For instance, :: can be now proven completely serializable.

It also adds some caching for CHA analysis, so simple types and HK types with instantiated type parameters can be proven serializable only once. This caching is imperfect because it will only emit errors for a concrete class once, instead of more times. I will fix this error in the future, for now we can live with this limitation.

Finally, it adds tests that confirm that the new analysis works.