Introduce temporary recognition of `deprecatedInheritance`
Those classes defined with `deprecatedInheritance` have the same effect as `assumeClosed`. Among other, tuples are annotated with this special deprecation warning because they were supposed to be final but they were not from the beginning. We respect this annotation temporarily, until we find a better solution not to recognize tuples as open class hierarchies without whitelisting it explicitly.