Make `capture` only allow identifiers
Before, the `capture` function could capture any stable path that is not considered lazy. However, this posed problems because captured variables with `capture` were not rewritten in the spore header and hence `this` references will be allowed in spore bodies.

As they cannot be allowed, one alternative was to rewrite the spore bodies if we find `capture` instances. However, this could potentially change the evaluation semantics of the program (removing `spore` from the block will produce a different result).

To work around this, the best alternative is to prevent the use of selects and only allow identifiers. This commit introduces this change with several tests.