Find better solution by not using `_` in the typetag
Using `_` in the typetag of `CanSerializa` strangely asks for the symbol of `TransitiveChecker` which obviously is part of the compiler plugin and is not in the classpath. This happens because the scala compiler is setting it as the owner of existential type `_`. Using `Any` solves the issue, for now. But this will come back. They never disappear. They follow me.