Fix brittle implementation of macro proxies
This feels good. This change removes the brittle interaction between normal spores and macro-generated spores that was used for the `Excluded` checks. It's based out of a @phaller idea. It does not remove the whiteboxity yet because that's a non-trivial change that because of limitations in implicit resolution for methods we may need to keep. I'll explore removing it in a follow-up commit.

Value discarding does not work very well with these changes, but this is not severe. See https://github.com/scalacenter/spores/issues/20.

Spores pickling tests are disabled because after we've changed the API, the pickler generation has to change. Concretely, the fact that excluded types were completely ignored. I do not have time to fix this now.

It does simplify a lot the spores implementation. Goodbye incidental complexity!