Fix spore type inference
Scala's type inference does not interact well with spores type members. When the types `C` and `E` are not explicit in the call site, the type inferences tries luck with `Nothing`. Trying luck with `Nothing` means that the capture members of a non explicit spores are fixed to be `Nothing`, no matter whether the macros are whitebox or blackbox.

The strategy to fix this situation is the following: generate spores with Nothing type members when the user is not interested in the spore type (this happens when the call site only asks for a function and not a spore, for instance) and generate the correct type members when she is making it explicit.

However, making this distinction is not easy. This implementation uses a trick in the macro implementation to hijack the compiler, get the global universe, get the macro typer and have access to the original tree that are not modified by the macro framework. The reason we have to do this is because `ctx.macroApplication` is typechecked (the tree structure has changed) and we have no access to the enclosingTree in a reliable way (only enclosingSymbol).

Once we have this, we map the typechecked tree and the untypechecked tree passed to the typer using positions. This gives us access to the tree enclosing the spore and allow us to check whether the user has: 1. explicitly asked for a spore with type parameters; 2. used a type ascription; 3. explicitly set the spore type in the lhs of a `ValDef`.

We cannot remove whiteboxity of spores if we want users to have ease when creating them (because otherwise `Captured` is always inferred). However, whiteboxity makes typecheck some situations (for instance, situation 3) if the expected type member is `Nothing`. In order to avoid this, we explicitly fail with an imitated typechecker error.

There are a lot of new examples that check that users won't have any problem using spores.